+++ partiql-value/src/lib.rs
@@ -535,17 +535,10 @@ impl Value {
     pub fn is_number(&self) -> bool {
         matches!(self, Value::Integer(_) | Value::Real(_) | Value::Decimal(_))
     }
-
-    #[inline]
-    /// Returns true if and only if Value is null or missing
-    pub fn is_null_or_missing(&self) -> bool {
-        matches!(self, Value::Missing | Value::Null)
-    }
-
     #[inline]
     /// Returns true if and only if Value is null or missing
     pub fn is_absent(&self) -> bool {
-        self.is_null_or_missing()
+        matches!(self, Value::Missing | Value::Null)
     }
 
     #[inline]